fix for compilation with GCC / Clang

+++ src/json/json.c
@@ -358,7 +358,7 @@ json_node *json_object_new(void)
 static char *json_member_name_canonicalize(const char *member_name)
 {
     size_t cmember_size = strlen(member_name) + 1;
-    int cmember_length = 0;
+    size_t cmember_length = 0;
     char *cmember_name = calloc(1, cmember_size);
 
     for (const char *p = member_name; *p; ++p) {
